Ensure "dropped" columns are excluded
When a column is dropped within a transaction, postgres simply marks the
column as dropped in the system catalogs. We need to avoid including any
dropped columns when creating history tables and trigger code (this can
occur if a table is altered then its history triggers are rebuilt in the
same transaction, for example).